refactor(exec): move ush_execute() into exec.c
