+++ include/exec.h
@@ -1,6 +1,8 @@
 #ifndef EXEC_H
 #define EXEC_H
 
+int ush_execute(char **args);
+
 int ush_launch(char **args);
 
 #endif /* EXEC_H */
