+++ src/exec/exec.c
@@ -1,0 +1,19 @@
+#include <string.h>
+#include "builtins.h"
+#include "exec.h"
+
+int ush_execute(char **args) {
+    int i;
+
+    if (args[0] == NULL) {
+        return 1;
+    }
+
+    for (i = 0; i < ush_num_builtins(); i++) {
+        if (strcmp(args[0], builtin_str[i]) == 0) {
+            return (*builtin_func[i])(args);
+        }
+    }
+
+    return ush_launch(args);
+}
