+++ src/shell.c
@@ -7,26 +7,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "builtins.h"
 #include "parser.h"
 #include "shell.h"
 #include "exec.h"
-
-//-----------------------------------------------------------------------------------------
-
-int ush_execute(char **args) {
-    int i;
-    if (args[0] == NULL) {
-        return 1;
-    }
-
-    for (i = 0; i < ush_num_builtins(); i++) {
-        if (strcmp(args[0], builtin_str[i]) == 0) {
-            return (*builtin_func[i])(args);
-        }
-    }
-    return ush_launch(args);
-}
 
 void ush_loop() {
     char *line;
